requirements.txt
accelerate==1.3.0
decord==0.6.0
deepspeed==0.16.2
diffusers==0.11.1
einops==0.8.1
huggingface_hub==0.30.2
imageio==2.37.0
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
peft==0.14.0
Pillow==11.2.1
qwen_vl_utils==0.0.11
Requests==2.32.3
safetensors==0.5.3
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.1
wandb==0.19.10
bert_score
pycocoevalcap
tensorboard
